Add partial_fit to scikit-learn wrappers of creme estimators #267
Also fix the Creme2SKLClassifier so it passes scikit-learn's checks.
Ensure SKL2CremeRegressor -> Creme2SKLRegressor is still a proper
estimator. Not ensured for classifiers at the moment due to discrepancy
in how classes are discovered (at first partial_fit for scikit-learn and
at __init__ for creme if prediction classes are to be consistent).