adding support for a variable vocab kmeans
Currently the kmeans model(s) require providing a set of indexed values to map
to the currently known centroids, and the models allow prediction given a feature
has not yet been seen, but it is not clear that it is easy to update models with new
features, and especially when they are strings! This pull request adds a new KMeans class,
VariableVocabKMeans that allows the user to provide a vocabulary (lookup/dict of keys as
words, values as counts or other value for the word) that is then mapped into the same
model space as the traditional Kmeans. I am also adding a function to get a centroid
and return the words instead of the indices.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>