updates to dbstream
currently we can hit a buggy case if we run predict before a learn, and
we do not have any labels. This change will fix that bug, ensuring that we only generate clusters from labels given that labels are defined. Secondly, on the merge step of a cluster / micro cluster, it is currently required to have perfectly overlapping features (which I do not think has to be the case?) so if we use a .get() instead of direct index we can default to a value of 0.0, setting the center value to 0 so the added value is 0.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>