dbstream: allow microcluster centoid update to have different terms
I was going to open an issue, but figured this was a small enough change that it would be easier to open a pull request to both discus and show what I mean! I was testing dbstream out more, and hit this case where the microcluster center couldn't update because of a key index error. This smells like the previous changes that we made to allow for updates only given the overlapping keys. So the solution of course is to add the "if j in x" to mapping of j to the new i center, and then you are allowed to have different keys without throwing an error.