refactoring neighbors models to use simple collections queue
The current implementation was done for a performance tweak, but it is limited because we cannot
be flexible to add slightly varying features, or to add additional metadata about a point in the window
such as a UID that can be used to look up a point later. This refactor takes the base logic from creme.neighbors,
and adds some additional features, 1. a minimum distance to consider adding a new datum to the window, ensuring
we have a more variable window that does better at learning, and 2. a class cleanup function for classification
to iterate over data in the window and ensure the classes known to the class accurately reflect the current
window, and 3. optional support to add a UID, which is useful in production cases where you do not just want
a prediction, you want to be able to look up the original datum using the identifer. I renamed the BaseNeighbors
class to be KNeighbors because it is useful for the user that wants to just return neighbors, as is, to do their
own processing. This is likely a start (and I have not run tests locally) so we can discuss further changes that
need to be done, a strategy for doing them, and what features / additional classes of the old implementation
we want to preserve

Signed-off-by: vsoch <vsoch@users.noreply.github.com>