Fix possible bug in hddm_w
I noticed what appears to be a bug in HDDM_W when two_sided_test is True. Because of how _has_mean_changed works, the check for a change when the mean decreases always defaults to True. I noticed that the recent refactoring removed two lines that, in essence, checked if sample 1 or sample 2 had just been initialized (i.e., lines 149 and 150:
if sample1.EWMA_estimator < 0 or sample2.EWMA_estimator < 0:
            return False
)

This pull request fixes that, if it indeed is a bug.