Change the regression target to 'float'
The Number interface matches all Python numbers and more. This includes
complex numbers. Most formulas are intended to deal with real numbers
only, and will error out if applied on ccomplex numbers.
Moreover, MyPy does not recognise the interfaces for the 'number'
module, and recommends to use Python's default numeric tower as
proposed in PEP 484.