CluStream: fix MC mismatch during prediction
Deals with a mismatch between centers used for distance calculation (mc.center) and centers used for cluster assignment (_mc_centers). This causes issues as _mc_centers is only updated when using k-Means. Due to cluster merging/deletion, the index found by the check for the closest micro-cluster may not necessarily refer to the same microcluster as the stored center in _mc_centers.