Full type hints for the module base (#1668)
* Remove the MyPy override for river.base

* Automaticaly infer simple type annotations

We use Autotyping to add annotations where they are obvious from the
context.

* Remove unused ignores

* Start adding type hints

* Correct a confusion in arguments list annotations

Arbitrary argument lists should be annotated with the type of the
arguments and not as a tuple containing the type; unless the argument
should all be tuples.

* Complete general annotations

* Add annotations to generics

* Make tags a set instead of a dict

* Use Iterator from the collections.abc module

typing.Iterator is deprecated since Python 3.9.

* Add some supplementary hints in other modules

* Check if the _multiclass attribute exists

The wrapped model can be any type of model, but only the classifiers
have a _multiclass property. Before acessing it, we must make sure the
attribute exists.

* The Grouper should accept all transformers

base.Transformer corresponds only to unbatched unsupervied transformers.
Other transformers should also be accepted for grouping.

* Change the regression target to 'float'

The Number interface matches all Python numbers and more. This includes
complex numbers. Most formulas are intended to deal with real numbers
only, and will error out if applied on ccomplex numbers.
Moreover, MyPy does not recognise the interfaces for the 'number'
module, and recommends to use Python's default numeric tower as
proposed in PEP 484.

* Annotate the kwargs argument for predict_*

* Make base.Ensemble and base.Wrapper generic

An ensemble accepts a single type of model that inherits from Estimator.
Wrapper lost the information about the wrapped model, keeping only the
fact it is an estimator.
In both cases, making the class generic allows retrieving the original
class of the wrapped model.

The nicer "type parameter" syntax was introduced only in 3.12, so I use
TypeVars to indicate genericity here in order to ensure compatibility.
They otherwise have the same meaning.

Wrapper now inherits for typing.Generic.
Before Python 3.12, generic classes must inherit from a generic class to
be considered generic. If the class introduces the genericity (it does
not inherit from a generic class), it must inherit from the Generic
class.
The Generic class does only one thing: mark its children as generic for
type checkers.

* Add type hints for some container variables

* Only import compose when type checking

* Introduce basic protocols for learning and predicting

These protocols only cover a few of the cases found in the codebase, but
they may be useful to start with?
I found it hard to give annotations that respect the substitution
principle.

* Export BaseTransformer

* Ignore errors in other modules

These error were indirectly caused by changes in `base`, where the
stricter typing invalidates the previously ambiguous (but valid) type
checks.
These errors can be treated in their own modules.

* Formatting

* Add a changelog note

There is one (hopefully minor) breaking change and two informative (no
change in behaviour) notes.