Define vectors and matrices for math
Vectors are mappings of any type to a float; matrices are mappings of
a 2-tuple (containing any type) to a float.

Using general protocols allows for covariance on the "value" generic
type, which is necesary if one wants to use the numeric pyramid (e.g.
input ints where floats are specified).